T-2.2: delete the dead ARSO/Vremenar subsystem from api.ts
Executes D-2. api.ts 1455 -> 637 lines (-845, +27).

Salvaged first, in this commit: rankPercentile and categorizeArso move
verbatim from api.ts:154-162,194-201 into a new percentile.ts, with the
ComputedPercentiles interface they need and CAT_COLORS (shared with the
live categorizeEra5). They are the correct empirical-rank implementation
D-6 requires and T-4.1 depends on. Deliberately NOT wired into the
display -- only CAT_COLORS is imported today -- so no displayed value
can move. api.ts -> percentile.ts is one-directional, so wiring them in
under T-4.1 cannot create a cycle.

Deleted: the SA/VR datasette+Vremenar bases, saGet, vremenarTemp,
vremenarIdMap, arsoStationIds, ARSO_NATIONAL, arsoStationId, the two
computeArso*Percentiles, fetchArsoAllDaily/Min and their caches,
_theilSen, _fitNBGLM, _normalCDF, _lgamma, _median, _mean,
_buildArsoRegressionResult, _buildArsoCalendar, arsoSeason,
percentileOf, fetchArsoSeasonHeatmap, fetchArsoTropical, fetchArsoTrend,
and every ARSO branch in fetchTodayStatus, fetchLast7, fetchDailyWindow,
fetchSeasonHeatmap, fetchRegression, fetchCalendar, fetchAnnualTrend and
fetchEra5Tropical.

Unreachability re-verified rather than assumed, on four independent
gates: fetchMeta mints only source:"era5" names; arsoStationIds is
declared [] and never assigned; ARSO_NATIONAL is never passed as a loc
(the UI's only national loc is ERA5_NATIONAL, and loc has one writer);
and tests/fixtures/index.json holds 2040 routes with zero arso/vremenar
entries, so under the T-1.2 shim any such call would have thrown.

The client-side Open-Meteo forecast path is independent and untouched:
openMeteoMax/openMeteoNationalMax read era5Coords, never the Vremenar
map, and vremenarTemp's only two callers were inside deleted branches.

Kept: isArsoLoc, because three live components still branch on it, and
removing those branches touches TodayCard.tsx:197, which D-11 freezes
for Phases 2-3. ArsoTropicalData, which despite its name is the return
shape of the live fetchEra5Tropical. monthDayToDoy, now unreferenced but
a named target of T-4.5 and cited by tests/snapshot/cases.json:115.

fetchMeta still requests _col=station_id although nothing reads it: the
fixture layer keys on the exact request URL, so shortening the query
would make the page's first request a fixture miss. Commented, to be
dropped at the Phase 4 re-record.

Comment-only updates to clock.ts (the new Date( audit re-run, 17 -> 13
hits; its two api.ts citations were the deleted ARSO branches) and to
fixtures/install.ts (the five swallowed-catch line numbers re-derived).

No published value moves. yarn snapshot:check: identical to
tests/fixtures/snapshot.json (18 cases, 2 stations, 0 fixture misses).
yarn build exits 0. Forced typecheck 91 -> 88 errors.
